🔨🤖 Inline shared.py into inequality_comparison
With poverty_inequality_file gone, inequality_comparison is the only consumer of these
helpers, so "shared" was a misnomer. Move the keyvars-assembly functions into the step
file (right after run()) and delete shared.py. Output unchanged.
